Unify user provider and trigger postAuthenticate on every login
