--- src/Security/User/BackendUserProvider.php
@@ -1,73 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-/*
- * This file is part of Contao.
- *
- * Copyright (c) 2005-2017 Leo Feyer
- *
- * @license LGPL-3.0+
- */
-
-namespace Contao\CoreBundle\Security\User;
-
-use Contao\BackendUser;
-use Contao\CoreBundle\Framework\ContaoFrameworkInterface;
-use Symfony\Component\Security\Core\Exception\UnsupportedUserException;
-use Symfony\Component\Security\Core\Exception\UsernameNotFoundException;
-use Symfony\Component\Security\Core\User\UserInterface;
-use Symfony\Component\Security\Core\User\UserProviderInterface;
-
-class BackendUserProvider implements UserProviderInterface
-{
-    /**
-     * @var ContaoFrameworkInterface
-     */
-    private $framework;
-
-    /**
-     * @param ContaoFrameworkInterface $framework
-     */
-    public function __construct(ContaoFrameworkInterface $framework)
-    {
-        $this->framework = $framework;
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function loadUserByUsername($username): UserInterface
-    {
-        $this->framework->initialize();
-
-        /** @var BackendUser $adapter */
-        $adapter = $this->framework->getAdapter(BackendUser::class);
-
-        if (($user = $adapter->loadUserByUsername($username)) instanceof BackendUser) {
-            return $user;
-        }
-
-        throw new UsernameNotFoundException(sprintf('Username "%s" does not exist.', $username));
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function refreshUser(UserInterface $user): UserInterface
-    {
-        if ($user instanceof BackendUser) {
-            return $this->loadUserByUsername($user->getUsername());
-        }
-
-        throw new UnsupportedUserException(sprintf('Instances of "%s" are not supported.', \get_class($user)));
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function supportsClass($class): bool
-    {
-        return BackendUser::class === $class;
-    }
-}
