--- src/Security/User/FrontendUserProvider.php
@@ -1,73 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-/*
- * This file is part of Contao.
- *
- * Copyright (c) 2005-2017 Leo Feyer
- *
- * @license LGPL-3.0+
- */
-
-namespace Contao\CoreBundle\Security\User;
-
-use Contao\CoreBundle\Framework\ContaoFrameworkInterface;
-use Contao\FrontendUser;
-use Symfony\Component\Security\Core\Exception\UnsupportedUserException;
-use Symfony\Component\Security\Core\Exception\UsernameNotFoundException;
-use Symfony\Component\Security\Core\User\UserInterface;
-use Symfony\Component\Security\Core\User\UserProviderInterface;
-
-class FrontendUserProvider implements UserProviderInterface
-{
-    /**
-     * @var ContaoFrameworkInterface
-     */
-    private $framework;
-
-    /**
-     * @param ContaoFrameworkInterface $framework
-     */
-    public function __construct(ContaoFrameworkInterface $framework)
-    {
-        $this->framework = $framework;
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function loadUserByUsername($username): UserInterface
-    {
-        $this->framework->initialize();
-
-        /** @var FrontendUser $adapter */
-        $adapter = $this->framework->getAdapter(FrontendUser::class);
-
-        if (($user = $adapter->loadUserByUsername($username)) instanceof FrontendUser) {
-            return $user;
-        }
-
-        throw new UsernameNotFoundException(sprintf('Username "%s" does not exist.', $username));
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function refreshUser(UserInterface $user): UserInterface
-    {
-        if ($user instanceof FrontendUser) {
-            return $this->loadUserByUsername($user->getUsername());
-        }
-
-        throw new UnsupportedUserException(sprintf('Instances of "%s" are not supported.', \get_class($user)));
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function supportsClass($class): bool
-    {
-        return FrontendUser::class === $class;
-    }
-}
